Fix connect sessions.
Before 0.4.0, sessions were stored in the DB as JSON strings.
As of 0.4.0, they were stored as objects rather than strings, but
the connect-session cookie.data getter results were stored, breaking
a connect-session assumption.

In this version, we use JSON.parse(JSON.stringify()) on the session
data to avoid storing the .data getter results.  We also detect
whether a session in the DB is a string, and if so parse it as JSON,
for backwards compatibility with pre-0.4.0 databases.